Refactor handler and service package names for consistency and clarity
