+++ internal/handler/urlExtractor/handler.go
@@ -1,4 +1,4 @@
-package url_extractor
+package urlExtractor
 
 import (
 	"net/http"
