+++ internal/handler/urlShortener/handler.go
@@ -1,4 +1,4 @@
-package url_shortener
+package urlShortener
 
 import (
 	"fmt"
