+++ internal/service/urlShortener/shorten_url_base62.go
@@ -1,4 +1,4 @@
-package url_shortener
+package urlShortener
 
 import (
 	"crypto/md5"
