+++ internal/service/urlShortener/shorten_url_base62_test.go
@@ -1,4 +1,4 @@
-package url_shortener
+package urlShortener
 
 import (
 	"math/big"
